Separate collection support from vocabulary support
**ISSUE**
We initially built the Collection UI behavior assuming that
collection labels would be handled as a kind of vocabulary.

While we want the UI for Collections and Vocabularies to be as
similar as possible if not identical; their differences
under-the-hood are too large to abrstract away right now.

In order to simplify implementing the Vocabulary UI, we're
splitting out the data_type and UI support for Collections, so the
two aren't fighting to overload the same methods.