Add form helper to render local vocabulary terms
Previously, we had used the terminology 'voczbulary' for
the list of collection names and had a corresponding FormBuilder
helper.

We want to separate code to support collecitons in the UI from
code to support vocabularies. This change adds a separate helper
that renders a selection list of the terms associated with a
specific vocabulary.