Store term IDs instead of keys in Resource records
**ISSUE**
Term keys are not guaranteed to be unique across vocabularies.
In order to unambiguously look up the label, we would need both
the Term key and the Vocabulary key or ID. We don't currently have
this stored at the Resource level; to derive it, we would need to
load the Blueprint for the item, find the corresponding Field, and
then look up the Vocabulary associated with the field.

**RESOLUTION**
Store the Term ID in the Resrouce record instead of the key. The ID
is guaranteed to be unique across all terms and does not require
a secondary key to resolve unambiguously.