[REFACTOR] Move solrization code into Field model
**ISSUE**
The #to_solr method in the `Resource` model needs to know about a
number of different `Field` methods to produce a desired Solr document.
We'd like to encapsulate that knowlege in the `Field` class.

**RESOLUTION**
Refactor the solrization call to be a single method call to the field
with the value to solrize as the method parameter. This simplifies
the interface between `Resource` and `Field` models.

**NOTE**
This change also extracts the Solr naming code to it's own concern
because both RuboCop and CodeClimate were complaining about the
overall calss length.