Add ingest support for vocabulary fields
**ISSUE**
When importing content from a foreign system, we don't necessarily
have access to internal database ids for vocabulary terms. Therefore,
we'd like to be able to import using term labels or keys.

**RESOLUTION**
This change adds a pre-processor to the imorter that scans vocabulary
fields and maps the contents to the corresponding term IDs before
attempting to create a respository item.