[FIX] Allow resources to be saved with vocabulary fields from forms
**ISSUE**
**1)** When editing a resource (item or collection) in the UI, the form
submits empty vocabulary fields as an array containing a zero length
string - e.g. if we have a vocabulary named 'resource_type', we would
get something like this:
```
params: {"item" => {"resource_type" => [""]}
```

**2)** When submitting a form, vocabulary IDs are submitted as strings
but our validation code only handled integers. When the validation
code ran in tests, the test passed because `2 == 2`, but from our
forms, we got values where `"2" != 2`.

**RESOLUTION**
Add tests that mirror form input and update the code to handle the
types of values that can be passed as part of a params hash.