[FIX] Duplicate source_fields only populate last field
**ISSUE**
When multiple fields have the same value for `source_field` only the
last field is populated, any earlier Fields defined with the same
source field are omitted on import.

**RESOUTION**
Don't allow multiple fields to use the same source field. If behavior
like this is required, use an external tool to modify the manifest
prior to importing it. Create a differently named field in the manifest
with a copy of the relevant metadata. Now, each uniquely named field
in the manifest can be mapped to a different field for import.