Log invalid vocabulary terms during ingests
**ISSUE**
The previous code silently discarded any terms in vocabulary fields
that could not be matched to a corresponding entry in the vocabulary.
This could result in data loss without users' knowleged.

**RESOLUTION**
Allow the assignment of any value(s) to vocabulary fields. Look up
the fields as keys, labels, or ids against the vocabulary during
validation AND convert them to the corresponding Term id before
saving.