[BUG] Handle invalid vocabulary key changes gracefully
**ISSUE**
When editing a vocabulary and entering an invalid key, or valid key
with other invalid data that will not validate, clicking the "cancel"
button attempted to route to a link generated using the unpersisted
key.

**FIX**
Instead of using the unpersisted, and possibly invalid, key;
check for the existence of the persisted value (i.e. #key_was
instead of #key) available through the ActiveModel interface.

For more details, see https://api.rubyonrails.org/classes/ActiveModel/Dirty.html