Enhance Vocabulary editing form
This change adds two usability features to the Vocabulary editor
(creating and editing vocabularies).

1. Dynamically displays the suggested vocabulary key used in the
URL for the vocabulary. Users can override the suggested value by
typing in the respective field.

2. Dynamically resizes the text area for the note input. The note
is able to span multiple lines if needed.

NOTE:
The fragment ` data: {lpignore:'true'} ` instructs LastPass to ignore
the form field (for LP autocomplete) to prevent errors due to both
LastPass and the application adding listeners to the field.